+++ HISTORY.rst
@@ -29,16 +29,17 @@ TODO
 0.4.1.dev0
 ----------
 
+Features
+^^^^^^^^
+
 * Import from affymetrix dataset
+* Rely on original affymetrix coordinate system to determine illumina top alleles
 * Search samples *aliases* while importing genotypes
 * Clearly state when creating samples (ignore samples if not defined in database)
 * Track sample aliases for ``original_id``
 * Import samples from file by providing *country* and *breeds* values as parameters
 * Security updates
 * Fix github Workflow
-
-Features
-^^^^^^^^
 
 0.4.0 (2021-06-18)
 ------------------